profile and refactor users
