+++ requirements.txt
@@ -1,0 +1,2 @@
+Pillow==7.1.2
+Django==3.0.6
