+++ whatnow/tasks/views/__init__.py
@@ -5,9 +5,11 @@ from users.models import Users
 
 def index(request):
     user = get_object_or_404(User, id=request.session.get('_auth_user_id'))
-    user = get_object_or_404(Users, email=user.email)
     request.session['user_id'] = user.id
-    request.session['user_type'] = user.user_type.user_type
+    print(request.session.get('user_id'))
+    user_type = get_object_or_404(Users, user=user)
+
+    request.session['user_type'] = user_type.user_type.user_type
     return render(request, 'index.html', {})
