+++ whatnow/users/templates/users/profile.html
@@ -10,4 +10,8 @@
     <a href="{% url 'tasks:tasks:list' %}" class="navigation"><b>Manage your tasks</b></a>
     <strong class="logo">WhatNow</strong>
 </header>
-<h1>{{user.username}}</h1>
+<div class="media">
+	<img width="300" src="{{ user.profile.image.url }}" alt="profile picture">
+	<h2>{{ user.username }}</h2>
+	<h2>{{ user.email }}</h2>
+</div>
