+++ whatnow/users/templates/users/register.html
@@ -3,6 +3,4 @@
     {{ form.as_p }}
     <button type="submit">Sign Up</button>
 </form>
-<small>
-	Already Have An Account? <a href="{% url 'login' %}">Login</a>
-</small>
+
