+++ whatnow/users/urls.py
@@ -1,8 +1,7 @@
 from django.urls import path
-from users.views import register, profile
+from users.views import register
 
 
 urlpatterns = [
-    path('', profile, name='profile'),
     path('', register, name='register'),
 ]
