+++ whatnow/users/views/__init__.py
@@ -1,0 +1,1 @@
+from .register import register, profile, index
