+++ whatnow/whatnow/settings.py
@@ -121,5 +121,8 @@ USE_TZ = True
 
 STATIC_URL = '/static/'
 
+MEDIA_ROOT = os.path.join(BASE_DIR, 'media')
+MEDIA_URL = '/media/'
+
 LOGIN_REDIRECT_URL = '/tasks'
 LOGIN_URL = '/'
