+++ whatnow/whatnow/urls.py
@@ -2,6 +2,9 @@ from django.contrib import admin
 from django.contrib.auth import views as auth_views
 from django.urls import path, include
 from users.views import profile
+from django.conf import settings
+from django.conf.urls.static import static
+
 
 urlpatterns = [
     path('tasks/', include('tasks.urls')),
@@ -11,3 +14,6 @@ urlpatterns = [
     path('logout/', auth_views.LogoutView.as_view(template_name='users/logout.html'), name='logout'),
     path('profile/', profile, name='profile'),
 ]
+
+if settings.DEBUG:
+    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
